requirements.txt
colorama~=0.3.9
torch==1.3
numpy~=1.16.0
ruamel.yaml>=0.15.87,<0.16.5
pandas>=0.23.4
tqdm~=4.28
scikit-learn~=0.20.3
paramiko==2.4.2
boto3>=1.9.100
ray==0.7.6
requests~=2.21.0
Flask~=1.0.0
tensorboardx-hparams==1.7.2
GitPython~=2.1.11
sru==2.1.9
pygments~=2.3.1
nltk~=3.4.1
dill>=0.3.1
gensim==3.7.3
pip~=19.1
awscli==1.16.202
transformers~=2.2.1
jinja2~=2.10.1
fastBPE==0.1.0
ninja~=1.9.0
